requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
anthropic>=0.52.0
python-multipart>=0.0.20
python-dotenv>=1.0.0
Pillow>=11.0.0
httpx>=0.28.0
